+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ==================
 
 * FindByText and FindByTextContaining now support simple quotes (')
+* Proper support for XPath on waitFor and watAndClick
 * Some dependencies updated
 
 2.0.1 / 2019-05-08